+++ modules/AttributeGroup/src/Http/Controllers/AttributeGroupController.php
@@ -5,6 +5,7 @@ namespace Modules\AttributeGroup\Http\Controllers;
 use App\Http\Controllers\Controller;
 use Modules\AttributeGroup\Contracts\AttributeGroupRepositoryInterface;
 use Modules\AttributeGroup\Http\Requests\AttributeGroupRequest;
+use Modules\AttributeGroup\Models\AttributeGroup;
 use Modules\Category\Contracts\CategoryRepositoryInterface;
 use Modules\Core\Responses\AjaxResponse;
 
@@ -16,40 +17,46 @@ class AttributeGroupController extends Controller
     {
         $this->attributeGroupRepo = $attributeGroupRepo;
     }
+
     public function index(CategoryRepositoryInterface $categoryRepo)
     {
+        $this->authorize('view', AttributeGroup::class);
         $categories = $categoryRepo->all();
         $attributeGroups = $this->attributeGroupRepo->getAll();
-        return view('AttributeGroup::index' , compact('attributeGroups' , 'categories'));
+        return view('AttributeGroup::index', compact('attributeGroups', 'categories'));
     }
 
     public function store(AttributeGroupRequest $request)
     {
+        $this->authorize('manage', AttributeGroup::class);
         $this->attributeGroupRepo->store($request->all());
         newFeedback();
         return back();
     }
 
-    public function edit($attributeGroupId , CategoryRepositoryInterface $categoryRepo)
+    public function edit($attributeGroupId, CategoryRepositoryInterface $categoryRepo)
     {
+        $this->authorize('manage', AttributeGroup::class);
         $categories = $categoryRepo->all();
         $attributeGroup = $this->attributeGroupRepo->findById($attributeGroupId);
-        return view('AttributeGroup::edit' , compact('attributeGroup' , 'categories'));
+        return view('AttributeGroup::edit', compact('attributeGroup', 'categories'));
     }
 
-    public function update(AttributeGroupRequest $request ,$attributeGroupId)
+    public function update(AttributeGroupRequest $request, $attributeGroupId)
     {
-        $this->attributeGroupRepo->update($attributeGroupId , $request->all());
+        $this->authorize('manage', AttributeGroup::class);
+        $this->attributeGroupRepo->update($attributeGroupId, $request->all());
         newFeedback();
         return to_route('panel.attributeGroups.index');
     }
 
     public function destroy($attributeGroupId)
     {
-       $attributeGroup =  $this->attributeGroupRepo->findById($attributeGroupId);
+        $this->authorize('manage', AttributeGroup::class);
+        $attributeGroup = $this->attributeGroupRepo->findById($attributeGroupId);
         $this->attributeGroupRepo->destroy($attributeGroupId);
         return AjaxResponse::success(
-            'گروه مشخصات'. $attributeGroup->name .' با موفقیت حذف شد.'
+            'گروه مشخصات' . $attributeGroup->name . ' با موفقیت حذف شد.'
         );
 
     }
