+++ modules/Coupon/src/Http/Controllers/CouponController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use Modules\Core\Responses\AjaxResponse;
 use Modules\Coupon\Contracts\CouponRepositoryInterface;
 use Modules\Coupon\Http\Requests\CouponRequest;
+use Modules\Coupon\Models\Coupon;
 use Modules\Coupon\Repositories\CouponRepo;
 use Modules\User\Contracts\UserRepositoryInterface;
 
@@ -20,18 +21,21 @@ class CouponController extends Controller
 
     public function index()
     {
+        $this->authorize('view' , Coupon::class);
         $coupons = $this->couponRepo->getAll();
         return view('Coupon::coupons.index', compact('coupons'));
     }
 
     public function create(UserRepositoryInterface $userRepo)
     {
+        $this->authorize('manage' , Coupon::class);
         $users = $userRepo->getAll();
         return view('Coupon::coupons.create' , compact('users'));
     }
 
     public function store(CouponRequest $request)
     {
+        $this->authorize('manage' , Coupon::class);
         $data = $request->all();
         $data['start_date'] = convertJalaliToDate($request->start_date , 'Y/m/d H:i' );
         $data['end_date'] = convertJalaliToDate($request->end_date , 'Y/m/d H:i' );
@@ -43,6 +47,7 @@ class CouponController extends Controller
 
     public function edit($couponId , UserRepositoryInterface $userRepo)
     {
+        $this->authorize('manage' , Coupon::class);
         $users = $userRepo->getAll();
         $coupon = $this->couponRepo->findById($couponId);
         return view('Coupon::coupons.edit', compact('coupon' , 'users'));
@@ -50,6 +55,7 @@ class CouponController extends Controller
 
     public function update(CouponRequest $request, $couponId)
     {
+        $this->authorize('manage' , Coupon::class);
         $data = $request->all();
         $data['start_date'] = convertJalaliToDate($request->start_date , 'Y/m/d H:i' );
         $data['end_date'] = convertJalaliToDate($request->end_date , 'Y/m/d H:i' );
@@ -61,6 +67,7 @@ class CouponController extends Controller
 
     public function destroy($couponId)
     {
+        $this->authorize('manage' , Coupon::class);
         $coupon = $this->couponRepo->findById($couponId);
 
         $this->couponRepo->destroy($couponId);
