+++ modules/Coupon/src/Resources/Views/common-discounts/index.blade.php
@@ -6,15 +6,17 @@
 @section('content')
     <div class="row">
         <div class="col-lg-12">
-            <div class="row mb-3">
-                <div class="col-sm-12 col-md-12 col-lg-2">
-                    <a href="{{ route('panel.commonDiscounts.create') }}"
-                       class="btn btn-primary
+            @if(auth()->user()->hasPermissionTo(\Modules\RolePermissions\Models\Permission::PERMISSION_MANAGE_COUPONS))
+                <div class="row mb-3">
+                    <div class="col-sm-12 col-md-12 col-lg-2">
+                        <a href="{{ route('panel.commonDiscounts.create') }}"
+                           class="btn btn-primary
                         btn-rounded waves-effect waves-light mb-2 me-2">
-                        <i class="mdi mdi-plus me-1"></i>
-                        @lang('Coupon::translation.commonDiscount.create')</a>
+                            <i class="mdi mdi-plus me-1"></i>
+                            @lang('Coupon::translation.commonDiscount.create')</a>
+                    </div>
                 </div>
-            </div>
+            @endif
             <div class="card">
                 <div class="card-body border border-5">
                     <div class="table-responsive">
@@ -29,7 +31,9 @@
                                 <th>تاریخ شروع</th>
                                 <th>تاریخ پایان</th>
                                 <th> وضعیت</th>
-                                <th> عملیات</th>
+                                @if(auth()->user()->hasPermissionTo(\Modules\RolePermissions\Models\Permission::PERMISSION_MANAGE_COUPONS))
+                                    <th> عملیات</th>
+                                @endif
                             </tr>
                             </thead>
                             <tbody>
@@ -46,23 +50,25 @@
                                         <span class="badge py-1 bg-{{ $discount->statusCssClass }}"> {{ $discount->status_name }}
                                         </span>
                                     </td>
-                                    <td>
-                                        <a class="btn btn-sm bg-transparent d-inline"
-                                           href="{{ route('panel.commonDiscounts.edit' , $discount->id) }}"><i
-                                                class="fa fa-pencil fa-15m text-success"></i></a>
+                                    @if(auth()->user()->hasPermissionTo(\Modules\RolePermissions\Models\Permission::PERMISSION_MANAGE_COUPONS))
+                                        <td>
+                                            <a class="btn btn-sm bg-transparent d-inline"
+                                               href="{{ route('panel.commonDiscounts.edit' , $discount->id) }}"><i
+                                                    class="fa fa-pencil fa-15m text-success"></i></a>
 
-                                        <a href="{{ route('panel.commonDiscounts.destroy' , $discount->id) }}"
-                                           onclick="deleteItem(event ,  '{{ route('panel.commonDiscounts.destroy' , $discount->id) }}')"
-                                           class="btn btn-sm bg-transparent d-inline delete-confirm"><i
-                                                class="fa fa-trash fa-15m text-danger"></i></a>
+                                            <a href="{{ route('panel.commonDiscounts.destroy' , $discount->id) }}"
+                                               onclick="deleteItem(event ,  '{{ route('panel.commonDiscounts.destroy' , $discount->id) }}')"
+                                               class="btn btn-sm bg-transparent d-inline delete-confirm"><i
+                                                    class="fa fa-trash fa-15m text-danger"></i></a>
 
-                                        <form action="{{ route('panel.commonDiscounts.destroy' , $discount->id) }}"
-                                              method="post"
-                                              id="destroy-brand-{{ $discount->id }}">
-                                            @csrf
-                                            @method('delete')
-                                        </form>
-                                    </td>
+                                            <form action="{{ route('panel.commonDiscounts.destroy' , $discount->id) }}"
+                                                  method="post"
+                                                  id="destroy-brand-{{ $discount->id }}">
+                                                @csrf
+                                                @method('delete')
+                                            </form>
+                                        </td>
+                                    @endif
                                 </tr>
                             @endforeach
                             </tbody>
